Add ESLint autofixes to the autoformatter
Adds in all of the plugins and parsers that Lucid
Software uses for our internal ESLint workflow.
_NOTE:_ this plugin will only work on projects
which utilize some subset of these plugins/parsers,
and if we want to support new plugins/parsers,
then we may need to build them into the formatting
extension. It is possible it may pick up on the
locally installed variants - but I'm less certain
about the context coming from an extension to be sure.

Also, when building this extension package, there are
warnings that we should start using a bundler:
"This extension consists of 5751 files, out of which 3334 are JavaScript
files. For performance reasons, you should bundle your extension:
https://aka.ms/vscode-bundle-extension . You should also exclude
unnecessary files by adding them to your .vscodeignore:
https://aka.ms/vscode-vscodeignore"